+++ tests/conftest.py
@@ -1,24 +1,11 @@
 # -*- coding: utf-8 -*-
 
 import datetime
-import json
-import os
 from typing import Dict, Generator, Union
 import pytest
 import requests
 
 from bookops_worldcat import WorldcatAccessToken, MetadataSession
-
-
-@pytest.fixture
-def live_keys() -> None:
-    if not os.getenv("GITHUB_ACTIONS"):
-        fh = os.path.expanduser("~/.oclc/nyp_wc_test.json")
-        with open(fh, "r") as file:
-            data = json.load(file)
-            os.environ["WCKey"] = data["key"]
-            os.environ["WCSecret"] = data["secret"]
-            os.environ["WCScopes"] = data["scopes"]
 
 
 @pytest.fixture
